Add visual representation for thesis and update game logic
Added a new python file main_visual.py to display a visual representation of the game state for the thesis. Enhanced the game logic in the parallel.ipynb and main.ipynb files to improve the heuristic function. Also, saved game states were adjusted for better understanding of the game progression.